reworked tests
